flask_site/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
google-cloud-storage==1.14.0
google-api-python-client>=1.6.7
oauth2client>=4.1.2
